github_events_api/pyproject.toml
[tool.poetry]
name = "github-events-api"
version = "0.1.0"
description = "Fetches Github events from database, aggregates and displays them."
authors = ["Richard Volcko <rikvolcko@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
toml = "^0.10.2"
python-dotenv = "^1.1.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
